Refactor all files to be under ipman package
