+++ src/main/java/ipman/commands/Command.java
@@ -1,4 +1,4 @@
-package commands;
+package ipman.commands;
 
 public interface Command {
     void execute(Context context);
