+++ src/main/java/ipman/commands/Context.java
@@ -1,0 +1,7 @@
+package ipman.commands;
+
+import ipman.models.TaskList;
+import ipman.models.TasksFileManager;
+import ipman.ui.Ui;
+
+public record Context(TaskList tasks, Ui ui, TasksFileManager fileManager) {}
