+++ src/main/java/ipman/commands/ExitException.java
@@ -1,3 +1,3 @@
-package commands;
+package ipman.commands;
 
 public class ExitException extends RuntimeException {}
