+++ src/main/java/ipman/models/SerializationException.java
@@ -1,4 +1,4 @@
-package models;
+package ipman.models;
 
 public class SerializationException extends RuntimeException {
     public SerializationException(String message) {
